centralize integration llm config, no-cache-dir
