+++ .github/workflows/e2e-registry1-weekly.yaml
@@ -58,7 +58,7 @@ jobs:
           python-version-file: "pyproject.toml"
 
       - name: Install Python Dependencies
-        run: pip install ".[dev]" "src/leapfrogai_api" "src/leapfrogai_sdk"
+        run: pip install ".[dev]" "src/leapfrogai_api" "src/leapfrogai_sdk" --no-cache-dir
 
       - name: Setup Playwright
         run: |
